+++ src/main/java/br/dev/andrestamatto/identityhub/infrastructure/repository/JpaUserRepositoryAdapter.java
@@ -1,0 +1,35 @@
+package br.dev.andrestamatto.identityhub.infrastructure.repository;
+
+import br.dev.andrestamatto.identityhub.application.ports.output.UserRepository;
+import br.dev.andrestamatto.identityhub.domain.entities.User;
+import br.dev.andrestamatto.identityhub.domain.valueobjects.Username;
+import br.dev.andrestamatto.identityhub.infrastructure.repository.entity.UserJpaEntity;
+import br.dev.andrestamatto.identityhub.infrastructure.repository.mapper.UserEntityMapper;
+import br.dev.andrestamatto.identityhub.infrastructure.repository.ports.SpringDataUserRepository;
+
+public class JpaUserRepositoryAdapter implements UserRepository {
+    private final SpringDataUserRepository repository;
+
+    public JpaUserRepositoryAdapter(SpringDataUserRepository repository) {
+        this.repository = repository;
+    }
+
+    @Override
+    public boolean existsBy(Username username) {
+        return repository.existsByUsername(username.value());
+    }
+
+    @Override
+    public User save(User user) {
+        UserJpaEntity entity = UserEntityMapper.jpaEntityFrom(user);
+        UserJpaEntity saved = repository.save(entity);
+        return UserEntityMapper.toDomain(saved);
+    }
+
+    @Override
+    public User findByUsername(Username username) {
+        return repository.findByUsername(username.value())
+                .map(UserEntityMapper::toDomain)
+                .orElse(null);
+    }
+}