+++ src/main/java/br/dev/andrestamatto/identityhub/infrastructure/security/SecurityConfiguration.java
@@ -1,0 +1,18 @@
+package br.dev.andrestamatto.identityhub.infrastructure.security;
+
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@Configuration
+@EnableConfigurationProperties(SecurityProperties.class)
+public class SecurityConfiguration {
+
+    @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+}