+++ src/main/java/br/dev/andrestamatto/identityhub/interfaces/rest/config/RestResponseConfiguration.java
@@ -1,0 +1,15 @@
+package br.dev.andrestamatto.identityhub.interfaces.rest.config;
+
+import br.dev.andrestamatto.identityhub.interfaces.rest.mapper.UserResponseMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RestResponseConfiguration {
+
+    @Bean
+    public UserResponseMapper registerUserResponseMapper() {
+        return new UserResponseMapper();
+    }
+
+}